feat: decode video in web worker
Moves the decoding of the video and the subsequent rendering to a canvas to a web worker. This way, the main thread is no longer blocked by these resource-heavy tasks, and the experience improves.

Since the task is moved to another thread, the component is ready sooner, showing the video as soon as it is loaded, and replacing it with the canvas version only ofter the frames have been extracted. The component should be interactive at all times.

Currently, easing is not yet implemented in the canvas version.